requirements.txt
Flask==1.1.2
Jinja2==2.11.2
MarkupSafe==1.1.1
PyJWT==1.7.1
PyYAML==5.3.1
Werkzeug==1.0.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
cryptography==3.1.1
firebase-admin==4.4.0
google-api-core==1.22.2
google-api-python-client==1.12.3
google-auth==1.22.0
google-auth-httplib2==0.0.4
google-cloud-core==1.4.1
google-cloud-firestore==1.9.0
google-cloud-storage==1.31.2
google-crc32c==1.0.0
google-resumable-media==1.0.0
googleapis-common-protos==1.52.0
grpcio==1.32.0
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
msal==1.5.0
oauthlib==3.1.0
pathvalidate==2.3.0
pip==20.2.3
pycparser==2.20
python-dateutil==2.8.1
requests==2.24.0
requests-oauthlib==1.3.0
sendgrid==6.4.7
setuptools==50.3.0
six==1.15.0
urllib3==1.25.10
websocket-client==0.48.0